Add JDBC tests for Spring Data

+++ src/main/java/com/pet/project/cinemacms/Models/AboutCinema/AboutCinema.java
@@ -113,11 +113,11 @@ public class AboutCinema {
         this.advertisings = advertisings;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
