+++ src/main/java/com/pet/project/cinemacms/Models/AboutCinema/Advertising.java
@@ -39,11 +39,11 @@ public class Advertising {
         this.aboutCinemaAdvert = aboutCinemaAdvert;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
