+++ src/main/java/com/pet/project/cinemacms/Models/Movie/Poster.java
@@ -66,11 +66,11 @@ public class Poster {
         this.movieTrailerUrl = movieTrailerUrl;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
